// Backend/src/types/type.response.ts
/* eslint-disable prettier/prettier */
// export type makesResponse = { name: string };
// export type modelsResponse = { model: string };
// export type getAllMakesResponse = { make: string };
// export type getMakeByIdResponse = { make: string };
// export type updateModelByIdResponse = { make: string; model: string };
export type getDistinctModel = { make: string; model: string };
export type vehicleResponse = {
  make: string;
  model: string;
  variant: string;
  body: string;
};